# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Profile",
    "description": "Profile API",
    "version": "1.0.0"
  },
  "host": "cse341-profile.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/region/{regionName}": {
      "get": {
        "summary": "Get users by region name",
        "parameters": [
          {
            "name": "regionName",
            "in": "path",
            "description": "Region name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users matching the region name"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/region/{regionName}": {
      "get": {
        "summary": "Get users by region",
        "parameters": [
          {
            "name": "regionName",
            "in": "path",
            "description": "Region name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users by region",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "404": {
            "description": "Region not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
  }